refactor(networks): make blockchainID and cChainBlockchainID mandatory in FlareNetwork interface
TICKET: WIN-7770
